Move command stuff to separate file "command.go"
