+++ internal/ws/request.go
@@ -2,17 +2,6 @@ package ws
 
 import "encoding/json"
 
-const (
-	CommandGetApiDevices      = Command("GET api.devices")
-	CommandPostApiDevice      = Command("POST api.device")
-	CommandPutApiDevice       = Command("PUT api.device")
-	CommandDeleteApiDevice    = Command("DELETE api.device")
-	CommandPostApiDevicePins  = Command("POST api.device.pins")
-	CommandPostApiDeviceColor = Command("POST api.device.color")
-)
-
-type Command string
-
 type Request struct {
 	Client  *Client `json:"-"`
 	Command Command `json:"command"`
